example/requirements.txt
Django==3.0.8
pillow==7.2.0
django-npm==1.0.0
-e ../

